pyproject.toml
[tool.poetry]
name = "test-aiogram-bot"
version = "0.1.0"
description = "Test task"
authors = ["Maxim <maksimovm117@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.10.0"
redis = "^5.0.8"
aioredis = "^2.0.1"
aiosqlite = "^0.20.0"
apscheduler = "^3.10.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
